requirements.txt
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
codecov==2.0.15
coverage==4.5.1
idna==2.6
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
pytest-cov==2.5.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
termcolor==1.1.0
urllib3==1.22
